refactor  numpy ndarray resolver as extension
