+++ _pydevd_bundle/pydevd_xml.py
@@ -124,12 +124,6 @@ if not sys.platform.startswith("java"):
                 pass #not available on all python versions
 
             try:
-                import numpy
-                type_map.append((numpy.ndarray, pydevd_resolver.ndarrayResolver))
-            except:
-                pass  #numpy may not be installed
-
-            try:
                 from django.utils.datastructures import MultiValueDict
                 type_map.insert(0, (MultiValueDict, pydevd_resolver.multiValueDictResolver))
                 #we should put it before dict
